refactor: extract excalidraw util modules
