+++ src/lib/components/ExcalidrawWrapper.svelte
@@ -1,7 +1,7 @@
 <script>
 	import { onMount, createEventDispatcher, tick } from 'svelte';
-	import { browser } from '$app/environment';
-	import { v4 as uuidv4 } from 'uuid';
+import { browser } from "$app/environment";
+import { createInitialImageElements, CANVAS_WIDTH, CANVAS_HEIGHT } from "$lib/utils/excalidrawTemplates.js";
 
 	export let data = null;
 	export let id = '';
@@ -19,11 +19,6 @@
 	let initialSceneData = null;
 	let excalidrawWrapper;
 	let hasInitialized = false;
-
-	const coneUrl = '/images/cone.webp';
-
-	const CANVAS_WIDTH = 500;
-	const CANVAS_HEIGHT = 600;
 
 	let fullscreenExcalidrawComponent;
 	let fullscreenContainer;
@@ -44,479 +39,6 @@
 		};
 	}
 
-	async function fetchImageAsDataURL(url) {
-		try {
-			// First check if image exists
-			const headResponse = await fetch(url, { method: 'HEAD' });
-			if (!headResponse.ok) {
-				console.error(`HEAD request failed with status ${headResponse.status} for URL: ${url}`);
-				return null;
-			}
-
-			// Then fetch the actual image
-			const getResponse = await fetch(url);
-			if (!getResponse.ok) {
-				console.error(`GET request failed with status ${getResponse.status} for URL: ${url}`);
-				return null;
-			}
-
-			const blob = await getResponse.blob();
-			return new Promise((resolve) => {
-				const reader = new FileReader();
-				reader.onloadend = () => {
-					resolve(reader.result);
-				};
-				reader.onerror = (error) => {
-					console.error(`Error converting ${url} to base64:`, error);
-					resolve(null);
-				};
-				reader.readAsDataURL(blob);
-			});
-		} catch (error) {
-			console.error(`Network error loading image from ${url}:`, error);
-			return null;
-		}
-	}
-
-	function createGuideRectangle() {
-		return {
-			id: uuidv4(),
-			type: 'rectangle',
-			x: 0,
-			y: 0,
-			width: CANVAS_WIDTH,
-			height: CANVAS_HEIGHT,
-			angle: 0,
-			strokeColor: '#ff0000',
-			backgroundColor: 'transparent',
-			fillStyle: 'solid',
-			strokeWidth: 2,
-			strokeStyle: 'dashed',
-			roughness: 0,
-			opacity: 30,
-			groupIds: [],
-			strokeSharpness: 'sharp',
-			seed: Math.floor(Math.random() * 1000000),
-			version: 1,
-			versionNonce: Math.floor(Math.random() * 1000000),
-			isDeleted: false,
-			locked: true
-		};
-	}
-
-	async function createInitialImageElements() {
-		const elements = [createGuideRectangle()];
-		const files = {};
-
-		// Add template-specific elements inside the guide rectangle
-		if (template === 'halfCourt') {
-			await addHalfCourtElements(elements, files);
-		} else if (template === 'fullCourt') {
-			await addFullCourtElements(elements, files);
-		}
-
-		// Always add the sidebar elements
-		await addSidebarElements(elements, files);
-
-		return { elements, files };
-	}
-
-	function createHoopSet(x, y) {
-		// Create a group ID for all hoops in this set
-		const hoopGroupId = uuidv4();
-		const hoopSizes = [
-			{ width: 40, height: 80 },
-			{ width: 40, height: 120 },
-			{ width: 40, height: 100 }
-		];
-		const spacing = 20;
-		const elements = [];
-
-		hoopSizes.forEach((size, i) => {
-			// Calculate the Y position for the hoop circle by working backwards from the base
-			const poleHeight = size.height - size.width;
-			const hoopY = y - poleHeight - size.width;
-
-			const hoopCircle = {
-				type: 'ellipse',
-				x: x + i * (size.width + spacing) - size.width / 2,
-				y: hoopY,
-				width: size.width,
-				height: size.width,
-				strokeColor: '#000000',
-				backgroundColor: 'transparent',
-				fillStyle: 'solid',
-				strokeWidth: 2,
-				roughness: 0,
-				opacity: 100,
-				strokeStyle: 'solid',
-				id: uuidv4(),
-				angle: 0,
-				groupIds: [hoopGroupId],
-				seed: Math.floor(Math.random() * 1000000),
-				version: 1,
-				versionNonce: Math.floor(Math.random() * 1000000),
-				isDeleted: false
-			};
-
-			const pole = {
-				type: 'line',
-				x: x + i * (size.width + spacing),
-				y: hoopY + size.width,
-				points: [
-					[0, 0],
-					[0, poleHeight]
-				],
-				strokeColor: '#000000',
-				backgroundColor: 'transparent',
-				fillStyle: 'solid',
-				strokeWidth: 2,
-				roughness: 0,
-				opacity: 100,
-				strokeStyle: 'solid',
-				id: uuidv4(),
-				width: 0,
-				height: poleHeight,
-				angle: 0,
-				groupIds: [hoopGroupId],
-				seed: Math.floor(Math.random() * 1000000),
-				version: 1,
-				versionNonce: Math.floor(Math.random() * 1000000),
-				isDeleted: false
-			};
-
-			elements.push(hoopCircle, pole);
-		});
-
-		return elements;
-	}
-
-	async function addHalfCourtElements(elements, files) {
-		// Move hoops down further
-		const hoopSet = createHoopSet(CANVAS_WIDTH / 2 - 40, 180); // Moved from 120 to 180
-		elements.push(...hoopSet);
-
-		// Adjust player positions relative to new hoop position
-		const playerPositions = [
-			{ x: CANVAS_WIDTH / 2, y: 200, type: 'k' }, // Keeper (right at hoops)
-			{ x: CANVAS_WIDTH / 2 - 60, y: 260, type: 'b1' }, // Left Beater - moved closer
-			{ x: CANVAS_WIDTH / 2 + 60, y: 260, type: 'b2' }, // Right Beater - moved closer
-			{ x: CANVAS_WIDTH / 2 - 100, y: 360, type: 'c1' }, // Left Chaser
-			{ x: CANVAS_WIDTH / 2, y: 360, type: 'c2' }, // Center Chaser
-			{ x: CANVAS_WIDTH / 2 + 100, y: 360, type: 'c3' } // Right Chaser
-		];
-
-		await addPlayersToField(elements, files, playerPositions);
-	}
-
-	async function addFullCourtElements(elements, files) {
-		// Add hoops at top and bottom, ensuring equal distance from boundaries
-		const topHoops = createHoopSet(CANVAS_WIDTH / 2 - 40, 120); // Moved up from 150
-		const bottomHoops = createHoopSet(CANVAS_WIDTH / 2 - 40, CANVAS_HEIGHT - 120); // Moved up from 150
-
-		elements.push(...topHoops, ...bottomHoops);
-
-		const playerPositions = [
-			// Team 1 (Blue team - top half)
-			{ x: CANVAS_WIDTH / 2, y: 140, type: 'k', team: 'blue' }, // Keeper closer to hoops
-			{ x: CANVAS_WIDTH / 2 - 50, y: 180, type: 'b1', team: 'blue' }, // Beaters closer together
-			{ x: CANVAS_WIDTH / 2 + 50, y: 180, type: 'b2', team: 'blue' },
-			{ x: CANVAS_WIDTH / 2 - 80, y: 240, type: 'c1', team: 'blue' }, // Chasers in formation
-			{ x: CANVAS_WIDTH / 2, y: 240, type: 'c2', team: 'blue' },
-			{ x: CANVAS_WIDTH / 2 + 80, y: 240, type: 'c3', team: 'blue' },
-
-			// Team 2 (Red team - bottom half)
-			{ x: CANVAS_WIDTH / 2, y: CANVAS_HEIGHT - 140, type: 'k', team: 'red' }, // Keeper closer to hoops
-			{ x: CANVAS_WIDTH / 2 - 50, y: CANVAS_HEIGHT - 180, type: 'b1', team: 'red' }, // Beaters closer together
-			{ x: CANVAS_WIDTH / 2 + 50, y: CANVAS_HEIGHT - 180, type: 'b2', team: 'red' },
-			{ x: CANVAS_WIDTH / 2 - 80, y: CANVAS_HEIGHT - 240, type: 'c1', team: 'red' }, // Chasers in formation
-			{ x: CANVAS_WIDTH / 2, y: CANVAS_HEIGHT - 240, type: 'c2', team: 'red' },
-			{ x: CANVAS_WIDTH / 2 + 80, y: CANVAS_HEIGHT - 240, type: 'c3', team: 'red' }
-		];
-
-		await addPlayersToField(elements, files, playerPositions);
-	}
-
-	async function addPlayersToField(elements, files, positions) {
-		for (const pos of positions) {
-			const imageId = uuidv4();
-			const teamColor = pos.team || 'blue';
-			const imagePath = `/images/icons/${teamColor}-player-${pos.type}.png`;
-
-			// Convert image to base64
-			const dataURL = await fetchImageAsDataURL(imagePath);
-			if (!dataURL) {
-				console.warn(`Failed to load image: ${imagePath}`);
-				continue;
-			}
-
-			elements.push({
-				id: imageId,
-				type: 'image',
-				x: pos.x,
-				y: pos.y,
-				width: 40,
-				height: 40,
-				angle: 0,
-				strokeColor: 'transparent',
-				backgroundColor: 'transparent',
-				fillStyle: 'hachure',
-				strokeWidth: 1,
-				strokeStyle: 'solid',
-				roughness: 0,
-				opacity: 100,
-				groupIds: [],
-				strokeSharpness: 'sharp',
-				seed: Math.floor(Math.random() * 1000000),
-				version: 1,
-				versionNonce: Math.floor(Math.random() * 1000000),
-				isDeleted: false,
-				scale: [1, 1],
-				fileId: imageId,
-				status: 'idle'
-			});
-
-			files[imageId] = {
-				id: imageId,
-				dataURL,
-				staticPath: imagePath,
-				mimeType: 'image/png',
-				created: Date.now(),
-				lastRetrieved: Date.now()
-			};
-		}
-	}
-
-	async function addSidebarElements(elements, files) {
-		// Add hoops - positioned to the right of the guide box
-		const hoopSizes = [
-			{ width: 40, height: 80 },
-			{ width: 40, height: 120 },
-			{ width: 40, height: 100 }
-		];
-		const spacing = 20;
-		const startX = CANVAS_WIDTH + 100;
-		const baseY = 100;
-
-		// Create a group ID for all hoops
-		const hoopGroupId = uuidv4();
-
-		hoopSizes.forEach((size, i) => {
-			// Calculate the Y position for the hoop circle by working backwards from the base
-			const poleHeight = size.height - size.width;
-			const hoopY = baseY - poleHeight - size.width;
-
-			const hoopCircle = {
-				type: 'ellipse',
-				x: startX + i * (size.width + spacing) - size.width / 2,
-				y: hoopY,
-				width: size.width,
-				height: size.width,
-				strokeColor: '#000000',
-				backgroundColor: 'transparent',
-				fillStyle: 'solid',
-				strokeWidth: 2,
-				roughness: 0,
-				opacity: 100,
-				strokeStyle: 'solid',
-				id: uuidv4(),
-				angle: 0,
-				groupIds: [hoopGroupId],
-				seed: Math.floor(Math.random() * 1000000),
-				version: 1,
-				versionNonce: Math.floor(Math.random() * 1000000),
-				isDeleted: false
-			};
-
-			const pole = {
-				type: 'line',
-				x: startX + i * (size.width + spacing),
-				y: hoopY + size.width,
-				points: [
-					[0, 0],
-					[0, poleHeight]
-				],
-				strokeColor: '#000000',
-				backgroundColor: 'transparent',
-				fillStyle: 'solid',
-				strokeWidth: 2,
-				roughness: 0,
-				opacity: 100,
-				strokeStyle: 'solid',
-				id: uuidv4(),
-				width: 0,
-				height: poleHeight,
-				angle: 0,
-				groupIds: [hoopGroupId],
-				seed: Math.floor(Math.random() * 1000000),
-				version: 1,
-				versionNonce: Math.floor(Math.random() * 1000000),
-				isDeleted: false
-			};
-
-			elements.push(hoopCircle, pole);
-		});
-
-		// Add player icons grid with base64 conversion
-		const iconSets = [
-			'b-and-w-player',
-			'blue-player',
-			'canada-player',
-			'red-black-player',
-			'red-player',
-			'ubc-player',
-			'y-and-b-player',
-			'yellow-arrow-player'
-		];
-
-		const positions = [
-			{ type: 'k', x: 0 },
-			{ type: 'c1', x: 60 },
-			{ type: 'c2', x: 120 },
-			{ type: 'c3', x: 180 },
-			{ type: 'b1', x: 240 },
-			{ type: 'b2', x: 300 },
-			{ type: 's', x: 360 }
-		];
-
-		let startY = 140;
-		const iconSize = 40;
-		const rowSpacing = 50;
-		const baseX = CANVAS_WIDTH + 90;
-
-		// Add all the player icons
-		for (let setIndex = 0; setIndex < iconSets.length; setIndex++) {
-			const currentY = startY + setIndex * rowSpacing;
-
-			for (const position of positions) {
-				const imageId = uuidv4();
-				const imagePath = `/images/icons/${iconSets[setIndex]}-${position.type}.png`;
-
-				// Convert image to base64
-				const dataURL = await fetchImageAsDataURL(imagePath);
-				if (!dataURL) {
-					console.warn(`Failed to load sidebar icon: ${imagePath}`);
-					continue;
-				}
-
-				elements.push({
-					id: imageId,
-					type: 'image',
-					x: baseX + position.x,
-					y: currentY,
-					width: iconSize,
-					height: iconSize,
-					angle: 0,
-					strokeColor: 'transparent',
-					backgroundColor: 'transparent',
-					fillStyle: 'hachure',
-					strokeWidth: 1,
-					strokeStyle: 'solid',
-					roughness: 0,
-					opacity: 100,
-					groupIds: [],
-					strokeSharpness: 'sharp',
-					seed: Math.floor(Math.random() * 1000000),
-					version: 1,
-					versionNonce: Math.floor(Math.random() * 1000000),
-					isDeleted: false,
-					scale: [1, 1],
-					fileId: imageId,
-					status: 'idle'
-				});
-
-				files[imageId] = {
-					id: imageId,
-					dataURL,
-					staticPath: imagePath,
-					mimeType: 'image/png',
-					created: Date.now(),
-					lastRetrieved: Date.now()
-				};
-			}
-		}
-
-		// Add the balls and cone with base64 conversion
-		const balls = [
-			{
-				url: '/images/icons/quaffle.png',
-				x: CANVAS_WIDTH + 70 + 30,
-				y: 115,
-				size: { width: 16, height: 20 }
-			},
-			{
-				url: '/images/icons/bludger.png',
-				x: CANVAS_WIDTH + 70 + 90,
-				y: 115,
-				size: { width: 16, height: 20 }
-			},
-			{
-				url: '/images/icons/bludger.png',
-				x: CANVAS_WIDTH + 70 + 150,
-				y: 115,
-				size: { width: 16, height: 20 }
-			},
-			{
-				url: '/images/icons/bludger.png',
-				x: CANVAS_WIDTH + 70 + 210,
-				y: 115,
-				size: { width: 16, height: 20 }
-			},
-			{
-				url: '/images/cone.webp',
-				x: CANVAS_WIDTH + 70 + 270,
-				y: 115,
-				size: { width: 20, height: 20 },
-				scale: [0.25, 0.25]
-			}
-		];
-
-		for (const ball of balls) {
-			const imageId = uuidv4();
-
-			// Convert image to base64
-			const dataURL = await fetchImageAsDataURL(ball.url);
-			if (!dataURL) {
-				console.warn(`Failed to load ball/cone: ${ball.url}`);
-				continue;
-			}
-
-			elements.push({
-				id: imageId,
-				type: 'image',
-				x: ball.x,
-				y: ball.y,
-				width: ball.size.width,
-				height: ball.size.height,
-				angle: 0,
-				strokeColor: 'transparent',
-				backgroundColor: 'transparent',
-				fillStyle: 'hachure',
-				strokeWidth: 1,
-				strokeStyle: 'solid',
-				roughness: 0,
-				opacity: 100,
-				groupIds: [],
-				strokeSharpness: 'sharp',
-				seed: Math.floor(Math.random() * 1000000),
-				version: 1,
-				versionNonce: Math.floor(Math.random() * 1000000),
-				isDeleted: false,
-				scale: ball.scale || [1, 1],
-				fileId: imageId,
-				status: 'idle'
-			});
-
-			files[imageId] = {
-				id: imageId,
-				dataURL,
-				staticPath: ball.url,
-				mimeType: ball.url.endsWith('.webp') ? 'image/webp' : 'image/png',
-				created: Date.now(),
-				lastRetrieved: Date.now()
-			};
-		}
-	}
 
 	function zoomToIncludeAllElements(api) {
 		if (!api) return;
@@ -804,7 +326,7 @@
 
 			// If there's no data or empty data, create from scratch; else fix and load existing data
 			if (!data || (data.elements && data.elements.length === 0)) {
-				initialSceneData = await createInitialImageElements();
+				initialSceneData = await createInitialImageElements(template);
 			} else {
 				const fixedElements = fixGuideRectanglePosition([...data.elements]);
 				initialSceneData = {
